Move mock_responses to a file shared with all tests
