+++ tests/test_fedora_distro_aliases.py
@@ -1,30 +1,11 @@
-import os
-import json
-from unittest.mock import patch, MagicMock
+from unittest.mock import patch
 from munch import Munch
 from fedora_distro_aliases import (
     Distro,
     bodhi_active_releases,
     get_distro_aliases,
 )
-
-here = os.path.abspath(os.path.dirname(__file__))
-datadir = os.path.join(here, "data")
-
-
-def mock_responses(files):
-    """
-    We can use this function to comfortably mock responses from saved files.
-    """
-    responses = []
-    for name in files:
-        path = os.path.join(datadir, name)
-        with open(path, "r") as fp:
-            data = json.load(fp)
-        response = MagicMock()
-        response.json.return_value = data
-        responses.append(response)
-    return responses
+from . import mock_responses
 
 
 def test_distro():
